refactor: Change structure of routes
route 구조변경.
- route() 메서드를 사용한 경로 처리 방식에서 모듈식 경로 처리 방식으로
  변경: express.Router 미들웨어 시스템 이용.
- 라우터를 모듈로 생성하고, 미들웨어 기능을 로드, 경로 정의, 마운트

- handler 삭제
- Router모듈들을 저장할 routes 디렉토리 생성
- auth 디렉토리 안에 getUserData 함수를 가진 파일 생성
