--- back/handler/signup.js
@@ -1,6 +1,0 @@
-/* --------------------------------- handler -------------------------------- */
-const verifyEmail = req => {};
-
-const addUser = req => {};
-
-module.exports = { verifyEmail, addUser };
