+++ back/routes/signup.js
@@ -1,0 +1,13 @@
+const express = require('express');
+
+const router = express.Router();
+
+router.use((req, res, next) => {
+  next();
+});
+
+router.get('/', () => {});
+
+router.post('/', () => {});
+
+module.exports = router;
