+++ back/server.js
@@ -1,24 +1,8 @@
-/* --------------------------------- require -------------------------------- */
 const express = require('express');
 const dotenv = require('dotenv');
 const cookieParser = require('cookie-parser');
 const cors = require('cors');
-const {
-  signIn,
-  verifyEmail,
-  addUser,
-  logOut,
-  getAllWords,
-  getVoca,
-  addVoca,
-  removeVoca,
-  getWords,
-  addWord,
-  editWord,
-  removeWord,
-} = require('./handler');
 
-/* ------------------------------------ use middleware ----------------------------------- */
 const app = express();
 // <클라이언트에서 서버로 요청이 가능하도록 하는 기본 설정>
 // client와 server간의 통신을 위해 JSON 형식을 다룰거기 때문에 JSON 미들웨어 설치
@@ -30,30 +14,22 @@ app.use(cors({ origin: '*', methods: ['GET', 'POST', 'PATCH', 'DELETE'], credent
 
 dotenv.config();
 
-/* -------------------------------- route -------------------------------- */
-// signin page
-app.post('/signin', signIn);
+/* -------------------------- require router module ------------------------- */
+const signin = require('./routes/signin');
+const signup = require('./routes/signup');
+const logout = require('./routes/logout');
+const vocalist = require('./routes/vocalist');
+const wordlist = require('./routes/wordlist');
+const game = require('./routes/game');
 
-// signup page
-app.get('/signup', verifyEmail);
-app.post('/signup', addUser);
-
-// vocalist page, wordlist page, game page
-app.route('/logout').get(logOut);
-
-// vocalist page
-app.get('/vocalist', getVoca);
-app.post('/vocalist', addVoca);
-app.delete('/vocalist/:id', removeVoca);
-
-// wordlist page
-app.get('/wordlist', getWords);
-app.post('/wordlist', addWord);
-app.patch('/wordlist/:id', editWord);
-app.delete('/wordlist/:id', removeWord);
-
-// game page
-app.get('/game', getAllWords);
+/* ------------------------------- load Router module ------------------------------- */
+// /routes -> 각 path의 라우터를 모듈로 생성.
+app.use('/signin', signin);
+app.use('/signup', signup);
+app.use('/logout', logout);
+app.use('/vocalist', vocalist);
+app.use('/wordlist', wordlist);
+app.use('/game', game);
 
 app.listen(process.env.PORT, () => {
   console.log(`server is on ${process.env.PORT}`);
